Fix and extend tests; something weird is still going on though
