+++ tests/v5/mqtt5_client_tests.c
@@ -28,7 +28,7 @@ static bool s_is_within_percentage_of(uint64_t expected_time, uint64_t actual_ti
     return fabs(actual_percent) <= percentage;
 }
 
-static int s_aws_mqtt5_mock_server_send_packet(
+int aws_mqtt5_mock_server_send_packet(
     struct aws_mqtt5_server_mock_connection_context *connection,
     enum aws_mqtt5_packet_type packet_type,
     void *packet) {
@@ -52,7 +52,7 @@ static int s_aws_mqtt5_mock_server_send_packet(
     return AWS_OP_SUCCESS;
 }
 
-static int s_aws_mqtt5_mock_server_handle_connect_always_succeed(
+int aws_mqtt5_mock_server_handle_connect_always_succeed(
     void *packet,
     struct aws_mqtt5_server_mock_connection_context *connection,
     void *user_data) {
@@ -64,7 +64,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_always_succeed(
 
     connack_view.reason_code = AWS_MQTT5_CRC_SUCCESS;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 static int s_aws_mqtt5_mock_server_handle_pingreq_always_respond(
@@ -74,7 +74,7 @@ static int s_aws_mqtt5_mock_server_handle_pingreq_always_respond(
     (void)packet;
     (void)user_data;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PINGRESP, NULL);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PINGRESP, NULL);
 }
 
 static int s_aws_mqtt5_mock_server_handle_disconnect(
@@ -97,16 +97,9 @@ void s_publish_received_callback(const struct aws_mqtt5_packet_publish_view *pub
     (void)user_data;
 }
 
-AWS_STATIC_STRING_FROM_LITERAL(s_client_id, "HelloWorld");
+AWS_STRING_FROM_LITERAL(g_default_client_id, "HelloWorld");
 
-struct mqtt5_client_test_options {
-    struct aws_mqtt5_client_topic_alias_options topic_aliasing_options;
-    struct aws_mqtt5_packet_connect_view connect_options;
-    struct aws_mqtt5_client_options client_options;
-    struct aws_mqtt5_mock_server_vtable server_function_table;
-};
-
-static void s_mqtt5_client_test_init_default_options(struct mqtt5_client_test_options *test_options) {
+void aws_mqtt5_client_test_init_default_options(struct mqtt5_client_test_options *test_options) {
 
     struct aws_mqtt5_client_topic_alias_options local_topic_aliasing_options = {
         .outbound_topic_alias_behavior = AWS_MQTT5_COTABT_DISABLED,
@@ -116,7 +109,7 @@ static void s_mqtt5_client_test_init_default_options(struct mqtt5_client_test_op
 
     struct aws_mqtt5_packet_connect_view local_connect_options = {
         .keep_alive_interval_seconds = 30,
-        .client_id = aws_byte_cursor_from_string(s_client_id),
+        .client_id = aws_byte_cursor_from_string(g_default_client_id),
         .clean_start = true,
     };
 
@@ -141,7 +134,7 @@ static void s_mqtt5_client_test_init_default_options(struct mqtt5_client_test_op
     struct aws_mqtt5_mock_server_vtable local_server_function_table = {
         .packet_handlers = {
             NULL,                                                   /* RESERVED = 0 */
-            &s_aws_mqtt5_mock_server_handle_connect_always_succeed, /* CONNECT */
+            &aws_mqtt5_mock_server_handle_connect_always_succeed,   /* CONNECT */
             NULL,                                                   /* CONNACK */
             NULL,                                                   /* PUBLISH */
             NULL,                                                   /* PUBACK */
@@ -167,7 +160,7 @@ static int s_aws_mqtt5_client_test_init_default_connect_storage(
 
     struct aws_mqtt5_packet_connect_view connect_view = {
         .keep_alive_interval_seconds = 30,
-        .client_id = aws_byte_cursor_from_string(s_client_id),
+        .client_id = aws_byte_cursor_from_string(g_default_client_id),
         .clean_start = true,
     };
 
@@ -232,7 +225,7 @@ static bool s_last_lifecycle_event_is_connected(void *arg) {
     return s_last_life_cycle_event_is(test_fixture, AWS_MQTT5_CLET_CONNECTION_SUCCESS);
 }
 
-static void s_wait_for_connected_lifecycle_event(struct aws_mqtt5_client_mock_test_fixture *test_context) {
+void aws_wait_for_connected_lifecycle_event(struct aws_mqtt5_client_mock_test_fixture *test_context) {
     aws_mutex_lock(&test_context->lock);
     aws_condition_variable_wait_pred(
         &test_context->signal, &test_context->lock, s_last_lifecycle_event_is_connected, test_context);
@@ -245,7 +238,7 @@ static bool s_last_lifecycle_event_is_stopped(void *arg) {
     return s_last_life_cycle_event_is(test_fixture, AWS_MQTT5_CLET_STOPPED);
 }
 
-static void s_wait_for_stopped_lifecycle_event(struct aws_mqtt5_client_mock_test_fixture *test_context) {
+void aws_wait_for_stopped_lifecycle_event(struct aws_mqtt5_client_mock_test_fixture *test_context) {
     aws_mutex_lock(&test_context->lock);
     aws_condition_variable_wait_pred(
         &test_context->signal, &test_context->lock, s_last_lifecycle_event_is_stopped, test_context);
@@ -317,7 +310,7 @@ static void s_wait_for_disconnect_completion(struct aws_mqtt5_client_mock_test_f
     aws_mutex_unlock(&test_context->lock);
 }
 
-static int s_verify_client_state_sequence(
+int aws_verify_client_state_sequence(
     struct aws_mqtt5_client_mock_test_fixture *test_context,
     enum aws_mqtt5_client_state *expected_states,
     size_t expected_states_count) {
@@ -361,7 +354,7 @@ static int s_verify_simple_lifecycle_event_sequence(
     return AWS_OP_SUCCESS;
 }
 
-static int s_verify_received_packet_sequence(
+int aws_verify_received_packet_sequence(
     struct aws_mqtt5_client_mock_test_fixture *test_context,
     struct aws_mqtt5_mock_server_packet_record *expected_packets,
     size_t expected_packets_count) {
@@ -400,7 +393,7 @@ static int s_mqtt5_client_direct_connect_success_fn(struct aws_allocator *alloca
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
         .client_options = &test_options.client_options,
@@ -413,7 +406,7 @@ static int s_mqtt5_client_direct_connect_success_fn(struct aws_allocator *alloca
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_disconnect_view disconnect_options = {
         .reason_code = AWS_MQTT5_DRC_DISCONNECT_WITH_WILL_MESSAGE,
@@ -426,7 +419,7 @@ static int s_mqtt5_client_direct_connect_success_fn(struct aws_allocator *alloca
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, &disconnect_options, &completion_options));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
     s_wait_for_disconnect_completion(&test_context);
     s_wait_for_mock_server_to_receive_disconnect_packet(&test_context);
 
@@ -457,7 +450,7 @@ static int s_mqtt5_client_direct_connect_success_fn(struct aws_allocator *alloca
         AWS_MCS_STOPPED,
     };
 
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     struct aws_mqtt5_packet_connect_storage expected_connect_storage;
     ASSERT_SUCCESS(s_aws_mqtt5_client_test_init_default_connect_storage(&expected_connect_storage, allocator));
@@ -477,7 +470,7 @@ static int s_mqtt5_client_direct_connect_success_fn(struct aws_allocator *alloca
         },
     };
     ASSERT_SUCCESS(
-        s_verify_received_packet_sequence(&test_context, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
+        aws_verify_received_packet_sequence(&test_context, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
 
     aws_mqtt5_packet_connect_storage_clean_up(&expected_connect_storage);
 
@@ -501,7 +494,7 @@ static int s_mqtt5_client_simple_failure_test_fn(
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
         .client_options = &test_options.client_options,
@@ -528,7 +521,7 @@ static int s_mqtt5_client_simple_failure_test_fn(
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
         {
@@ -546,7 +539,7 @@ static int s_mqtt5_client_simple_failure_test_fn(
         AWS_MCS_CONNECTING,
         AWS_MCS_PENDING_RECONNECT,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -747,7 +740,7 @@ static int s_mqtt5_client_websocket_connect_handshake_failure_fn(struct aws_allo
 
 AWS_TEST_CASE(mqtt5_client_websocket_connect_handshake_failure, s_mqtt5_client_websocket_connect_handshake_failure_fn)
 
-static int s_aws_mqtt5_mock_server_handle_connect_always_fail(
+int aws_mqtt5_mock_server_handle_connect_always_fail(
     void *packet,
     struct aws_mqtt5_server_mock_connection_context *connection,
     void *user_data) {
@@ -759,7 +752,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_always_fail(
 
     connack_view.reason_code = AWS_MQTT5_CRC_BANNED;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 /* Connection failure test where overall connection fails due to a CONNACK error code */
@@ -769,10 +762,10 @@ static int s_mqtt5_client_direct_connect_connack_refusal_fn(struct aws_allocator
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
-        s_aws_mqtt5_mock_server_handle_connect_always_fail;
+        aws_mqtt5_mock_server_handle_connect_always_fail;
 
     struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
         .client_options = &test_options.client_options,
@@ -790,7 +783,7 @@ static int s_mqtt5_client_direct_connect_connack_refusal_fn(struct aws_allocator
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
         {
@@ -809,7 +802,7 @@ static int s_mqtt5_client_direct_connect_connack_refusal_fn(struct aws_allocator
         AWS_MCS_MQTT_CONNECT,
         AWS_MCS_CHANNEL_SHUTDOWN,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -826,7 +819,7 @@ static int s_mqtt5_client_direct_connect_connack_timeout_fn(struct aws_allocator
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* fast CONNACK timeout and don't response to the CONNECT packet */
     test_options.client_options.connack_timeout_ms = 2000;
@@ -848,7 +841,7 @@ static int s_mqtt5_client_direct_connect_connack_timeout_fn(struct aws_allocator
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
         {
@@ -867,7 +860,7 @@ static int s_mqtt5_client_direct_connect_connack_timeout_fn(struct aws_allocator
         AWS_MCS_MQTT_CONNECT,
         AWS_MCS_CHANNEL_SHUTDOWN,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -898,7 +891,7 @@ static void s_server_disconnect_service_fn(
         .reason_code = AWS_MQTT5_DRC_PACKET_TOO_LARGE,
     };
 
-    s_aws_mqtt5_mock_server_send_packet(mock_server, AWS_MQTT5_PT_DISCONNECT, &disconnect);
+    aws_mqtt5_mock_server_send_packet(mock_server, AWS_MQTT5_PT_DISCONNECT, &disconnect);
 }
 
 static int s_aws_mqtt5_server_disconnect_on_connect(
@@ -910,7 +903,7 @@ static int s_aws_mqtt5_server_disconnect_on_connect(
      * We intercept the CONNECT in order to correctly set the connack_sent test state.  Otherwise we risk sometimes
      * sending the DISCONNECT before the CONNACK
      */
-    int result = s_aws_mqtt5_mock_server_handle_connect_always_succeed(packet, connection, user_data);
+    int result = aws_mqtt5_mock_server_handle_connect_always_succeed(packet, connection, user_data);
 
     struct aws_mqtt5_server_disconnect_test_context *test_context = user_data;
     test_context->connack_sent = true;
@@ -925,7 +918,7 @@ static int s_mqtt5_client_direct_connect_from_server_disconnect_fn(struct aws_al
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* mock server sends a DISCONNECT packet back to the client after a successful CONNECTION establishment */
     test_options.server_function_table.service_task_fn = s_server_disconnect_service_fn;
@@ -953,7 +946,7 @@ static int s_mqtt5_client_direct_connect_from_server_disconnect_fn(struct aws_al
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
         {
@@ -976,7 +969,7 @@ static int s_mqtt5_client_direct_connect_from_server_disconnect_fn(struct aws_al
         AWS_MCS_CONNECTED,
         AWS_MCS_CHANNEL_SHUTDOWN,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -1059,7 +1052,7 @@ static int s_aws_mqtt5_client_test_init_ping_test_connect_storage(
     struct aws_mqtt5_packet_connect_view connect_view = {
         .keep_alive_interval_seconds =
             (uint16_t)aws_timestamp_convert(TEST_PING_INTERVAL_MS, AWS_TIMESTAMP_MILLIS, AWS_TIMESTAMP_SECS, NULL),
-        .client_id = aws_byte_cursor_from_string(s_client_id),
+        .client_id = aws_byte_cursor_from_string(g_default_client_id),
         .clean_start = true,
     };
 
@@ -1078,7 +1071,7 @@ static int s_mqtt5_client_ping_sequence_fn(struct aws_allocator *allocator, void
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* fast keep alive in order keep tests reasonably short */
     uint16_t keep_alive_seconds =
@@ -1105,7 +1098,7 @@ static int s_mqtt5_client_ping_sequence_fn(struct aws_allocator *allocator, void
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
     s_wait_for_n_pingreqs(&ping_context);
 
     /*
@@ -1127,7 +1120,7 @@ static int s_mqtt5_client_ping_sequence_fn(struct aws_allocator *allocator, void
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, &disconnect_view, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
     s_wait_for_mock_server_to_receive_disconnect_packet(&test_context);
 
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
@@ -1156,7 +1149,7 @@ static int s_mqtt5_client_ping_sequence_fn(struct aws_allocator *allocator, void
         AWS_MCS_CHANNEL_SHUTDOWN,
         AWS_MCS_STOPPED,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     struct aws_mqtt5_packet_connect_storage expected_connect_storage;
     ASSERT_SUCCESS(s_aws_mqtt5_client_test_init_ping_test_connect_storage(&expected_connect_storage, allocator));
@@ -1190,7 +1183,7 @@ static int s_mqtt5_client_ping_sequence_fn(struct aws_allocator *allocator, void
         },
     };
     ASSERT_SUCCESS(
-        s_verify_received_packet_sequence(&test_context, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
+        aws_verify_received_packet_sequence(&test_context, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
 
     ASSERT_SUCCESS(s_verify_ping_sequence_timing(&test_context));
 
@@ -1268,7 +1261,7 @@ static int s_mqtt5_client_ping_timeout_fn(struct aws_allocator *allocator, void 
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* fast keep alive in order keep tests reasonably short */
     uint16_t keep_alive_seconds =
@@ -1292,12 +1285,12 @@ static int s_mqtt5_client_ping_timeout_fn(struct aws_allocator *allocator, void 
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
     s_wait_for_disconnection_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
         {
@@ -1326,7 +1319,7 @@ static int s_mqtt5_client_ping_timeout_fn(struct aws_allocator *allocator, void 
         AWS_MCS_CLEAN_DISCONNECT,
         AWS_MCS_CHANNEL_SHUTDOWN,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -1336,49 +1329,47 @@ static int s_mqtt5_client_ping_timeout_fn(struct aws_allocator *allocator, void 
 
 AWS_TEST_CASE(mqtt5_client_ping_timeout, s_mqtt5_client_ping_timeout_fn)
 
-struct aws_connection_failure_wait_context {
-    size_t number_of_failures;
+struct aws_lifecycle_event_wait_context {
+    enum aws_mqtt5_client_lifecycle_event_type type;
+    size_t count;
     struct aws_mqtt5_client_mock_test_fixture *test_fixture;
 };
 
-static bool s_received_at_least_n_connection_failures(void *arg) {
-    struct aws_connection_failure_wait_context *context = arg;
+static bool s_received_at_least_n_events(void *arg) {
+    struct aws_lifecycle_event_wait_context *context = arg;
     struct aws_mqtt5_client_mock_test_fixture *test_fixture = context->test_fixture;
 
-    size_t failure_count = 0;
+    size_t actual_count = 0;
     size_t event_count = aws_array_list_length(&test_fixture->lifecycle_events);
     for (size_t i = 0; i < event_count; ++i) {
         struct aws_mqtt5_lifecycle_event_record *record = NULL;
         aws_array_list_get_at(&test_fixture->lifecycle_events, &record, i);
 
-        if (record->event.event_type == AWS_MQTT5_CLET_CONNECTION_FAILURE) {
-            failure_count++;
+        if (record->event.event_type == context->type) {
+            actual_count++;
         }
     }
 
-    return failure_count >= context->number_of_failures;
+    return actual_count >= context->count;
 }
 
-static void s_wait_for_n_connection_failure_lifecycle_events(
+void aws_mqtt5_wait_for_n_lifecycle_events(
     struct aws_mqtt5_client_mock_test_fixture *test_context,
-    size_t failure_count) {
-    struct aws_connection_failure_wait_context context = {
-        .number_of_failures = failure_count,
+    enum aws_mqtt5_client_lifecycle_event_type type,
+    size_t count) {
+    struct aws_lifecycle_event_wait_context context = {
+        .type = type,
+        .count = count,
         .test_fixture = test_context,
     };
 
     aws_mutex_lock(&test_context->lock);
     aws_condition_variable_wait_pred(
-        &test_context->signal, &test_context->lock, s_received_at_least_n_connection_failures, &context);
+        &test_context->signal, &test_context->lock, s_received_at_least_n_events, &context);
     aws_mutex_unlock(&test_context->lock);
 }
 
-#define RECONNECT_TEST_MIN_BACKOFF 500
-#define RECONNECT_TEST_MAX_BACKOFF 5000
-#define RECONNECT_TEST_BACKOFF_RESET_DELAY 5000
-
-static int s_verify_reconnection_exponential_backoff_timestamps(
-    struct aws_mqtt5_client_mock_test_fixture *test_fixture) {
+int aws_verify_reconnection_exponential_backoff_timestamps(struct aws_mqtt5_client_mock_test_fixture *test_fixture) {
     aws_mutex_lock(&test_fixture->lock);
 
     size_t event_count = aws_array_list_length(&test_fixture->lifecycle_events);
@@ -1421,7 +1412,7 @@ static int s_mqtt5_client_reconnect_failure_backoff_fn(struct aws_allocator *all
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* backoff delay sequence: 500, 1000, 2000, 4000, 5000, ... */
     test_options.client_options.retry_jitter_mode = AWS_EXPONENTIAL_BACKOFF_JITTER_NONE;
@@ -1430,7 +1421,7 @@ static int s_mqtt5_client_reconnect_failure_backoff_fn(struct aws_allocator *all
     test_options.client_options.min_connected_time_to_reset_reconnect_delay_ms = RECONNECT_TEST_BACKOFF_RESET_DELAY;
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
-        s_aws_mqtt5_mock_server_handle_connect_always_fail;
+        aws_mqtt5_mock_server_handle_connect_always_fail;
 
     struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
         .client_options = &test_options.client_options,
@@ -1444,11 +1435,11 @@ static int s_mqtt5_client_reconnect_failure_backoff_fn(struct aws_allocator *all
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_n_connection_failure_lifecycle_events(&test_context, 6);
+    aws_mqtt5_wait_for_n_lifecycle_events(&test_context, AWS_MQTT5_CLET_CONNECTION_FAILURE, 6);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* 6 (connecting, connection failure) pairs */
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
@@ -1498,7 +1489,7 @@ static int s_mqtt5_client_reconnect_failure_backoff_fn(struct aws_allocator *all
     ASSERT_SUCCESS(
         s_verify_simple_lifecycle_event_sequence(&test_context, expected_events, AWS_ARRAY_SIZE(expected_events)));
 
-    ASSERT_SUCCESS(s_verify_reconnection_exponential_backoff_timestamps(&test_context));
+    ASSERT_SUCCESS(aws_verify_reconnection_exponential_backoff_timestamps(&test_context));
 
     /* 6 (connecting, mqtt_connect, channel_shutdown, pending_reconnect) tuples (minus the final pending_reconnect) */
     enum aws_mqtt5_client_state expected_states[] = {
@@ -1509,7 +1500,7 @@ static int s_mqtt5_client_reconnect_failure_backoff_fn(struct aws_allocator *all
         AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN, AWS_MCS_PENDING_RECONNECT,
         AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
 
@@ -1547,7 +1538,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_succeed_on_nth(
 
     ++context->connection_attempts;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 static void s_aws_mqtt5_mock_server_disconnect_after_n_ms(
@@ -1574,7 +1565,7 @@ static void s_aws_mqtt5_mock_server_disconnect_after_n_ms(
             .reason_code = AWS_MQTT5_DRC_PACKET_TOO_LARGE,
         };
 
-        s_aws_mqtt5_mock_server_send_packet(mock_server, AWS_MQTT5_PT_DISCONNECT, &disconnect);
+        aws_mqtt5_mock_server_send_packet(mock_server, AWS_MQTT5_PT_DISCONNECT, &disconnect);
         context->connect_timestamp = 0;
     }
 }
@@ -1631,7 +1622,7 @@ static int s_mqtt5_client_reconnect_backoff_insufficient_reset_fn(struct aws_all
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     struct aws_mqtt5_mock_server_reconnect_state mock_server_state = {
         .required_connection_failure_count = 6,
@@ -1662,11 +1653,11 @@ static int s_mqtt5_client_reconnect_backoff_insufficient_reset_fn(struct aws_all
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_n_connection_failure_lifecycle_events(&test_context, 7);
+    aws_mqtt5_wait_for_n_lifecycle_events(&test_context, AWS_MQTT5_CLET_CONNECTION_FAILURE, 7);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* 6 (connecting, connection failure) pairs, followed by a successful connection, then a disconnect and reconnect */
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
@@ -1751,7 +1742,7 @@ static int s_mqtt5_client_reconnect_backoff_sufficient_reset_fn(struct aws_alloc
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     struct aws_mqtt5_mock_server_reconnect_state mock_server_state = {
         .required_connection_failure_count = 6,
@@ -1782,11 +1773,11 @@ static int s_mqtt5_client_reconnect_backoff_sufficient_reset_fn(struct aws_alloc
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_n_connection_failure_lifecycle_events(&test_context, 7);
+    aws_mqtt5_wait_for_n_lifecycle_events(&test_context, AWS_MQTT5_CLET_CONNECTION_FAILURE, 7);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* 6 (connecting, connection failure) pairs, followed by a successful connection, then a disconnect and reconnect */
     struct aws_mqtt5_client_lifecycle_event expected_events[] = {
@@ -1925,7 +1916,7 @@ static int s_aws_mqtt5_server_send_suback_on_subscribe(
         .reason_codes = s_suback_reason_codes,
     };
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_SUBACK, &suback_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_SUBACK, &suback_view);
 }
 
 /* Connection test where we succeed, send a SUBSCRIBE, and wait for a SUBACK */
@@ -1935,7 +1926,7 @@ static int s_mqtt5_client_subscribe_success_fn(struct aws_allocator *allocator, 
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_SUBSCRIBE] =
         s_aws_mqtt5_server_send_suback_on_subscribe;
@@ -1958,7 +1949,7 @@ static int s_mqtt5_client_subscribe_success_fn(struct aws_allocator *allocator, 
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_subscribe_view subscribe_view = {
         .subscriptions = s_subscriptions,
@@ -1975,7 +1966,7 @@ static int s_mqtt5_client_subscribe_success_fn(struct aws_allocator *allocator, 
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -2000,7 +1991,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_succeed_maximum_packet_size(
     connack_view.reason_code = AWS_MQTT5_CRC_SUCCESS;
     connack_view.maximum_packet_size = &maximum_packet_size;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 void s_aws_mqtt5_subscribe_complete_packet_size_too_small_fn(
@@ -2025,7 +2016,7 @@ static int s_mqtt5_client_subscribe_fail_packet_too_big_fn(struct aws_allocator 
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
         s_aws_mqtt5_mock_server_handle_connect_succeed_maximum_packet_size;
@@ -2048,7 +2039,7 @@ static int s_mqtt5_client_subscribe_fail_packet_too_big_fn(struct aws_allocator 
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_subscribe_view subscribe_view = {
         .subscriptions = s_subscriptions,
@@ -2065,7 +2056,7 @@ static int s_mqtt5_client_subscribe_fail_packet_too_big_fn(struct aws_allocator 
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -2087,7 +2078,7 @@ static int s_mqtt5_client_disconnect_fail_packet_too_big_fn(struct aws_allocator
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
         s_aws_mqtt5_mock_server_handle_connect_succeed_maximum_packet_size;
@@ -2110,7 +2101,7 @@ static int s_mqtt5_client_disconnect_fail_packet_too_big_fn(struct aws_allocator
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_byte_cursor long_reason_string_cursor = aws_byte_cursor_from_c_str(
         "Not valid because it includes the 0-terminator but we don't check utf-8 so who cares");
@@ -2127,7 +2118,7 @@ static int s_mqtt5_client_disconnect_fail_packet_too_big_fn(struct aws_allocator
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, &disconnect_view, &completion_options));
 
     s_wait_for_disconnect_completion(&test_context);
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -2157,7 +2148,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_succeed_receive_maximum(
     connack_view.reason_code = AWS_MQTT5_CRC_SUCCESS;
     connack_view.receive_maximum = &receive_maximum;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 struct send_puback_task {
@@ -2185,7 +2176,7 @@ void send_puback_fn(struct aws_task *task, void *arg, enum aws_task_status statu
         .packet_id = puback_response_task->packet_id,
     };
 
-    s_aws_mqtt5_mock_server_send_packet(puback_response_task->connection, AWS_MQTT5_PT_PUBACK, &puback_view);
+    aws_mqtt5_mock_server_send_packet(puback_response_task->connection, AWS_MQTT5_PT_PUBACK, &puback_view);
 
 done:
 
@@ -2281,7 +2272,7 @@ static int s_mqtt5_client_flow_control_receive_maximum_fn(struct aws_allocator *
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* send delayed pubacks */
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
@@ -2309,7 +2300,7 @@ static int s_mqtt5_client_flow_control_receive_maximum_fn(struct aws_allocator *
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     /* send a bunch of publishes */
     for (size_t i = 0; i < RECEIVE_MAXIMUM_PUBLISH_COUNT; ++i) {
@@ -2339,7 +2330,7 @@ static int s_mqtt5_client_flow_control_receive_maximum_fn(struct aws_allocator *
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /*
      * verify that the maximum number of in-progress qos1 publishes on the server was never more than what the
@@ -2434,7 +2425,7 @@ static int s_mqtt5_client_publish_timeout_fn(struct aws_allocator *allocator, vo
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
         s_aws_mqtt5_mock_server_handle_timeout_publish;
@@ -2454,7 +2445,7 @@ static int s_mqtt5_client_publish_timeout_fn(struct aws_allocator *allocator, vo
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_publish_completion_options completion_options = {
         .completion_callback = &s_publish_timeout_publish_completion_fn,
@@ -2488,7 +2479,7 @@ static int s_mqtt5_client_publish_timeout_fn(struct aws_allocator *allocator, vo
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -2514,7 +2505,7 @@ static int s_aws_mqtt5_mock_server_handle_publish_puback(
         .packet_id = publish_view->packet_id,
     };
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view);
 }
 
 #define IOT_CORE_THROUGHPUT_PACKETS 21
@@ -2524,7 +2515,7 @@ static uint8_t s_large_packet_payload[127 * 1024];
 static int s_do_iot_core_throughput_test(struct aws_allocator *allocator, bool use_iot_core_limits) {
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* send pubacks */
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
@@ -2553,7 +2544,7 @@ static int s_do_iot_core_throughput_test(struct aws_allocator *allocator, bool u
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     /* send a bunch of large publishes */
     aws_secure_zero(s_large_packet_payload, AWS_ARRAY_SIZE(s_large_packet_payload));
@@ -2586,7 +2577,7 @@ static int s_do_iot_core_throughput_test(struct aws_allocator *allocator, bool u
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
 
@@ -2627,7 +2618,7 @@ AWS_TEST_CASE(mqtt5_client_flow_control_iot_core_throughput, s_mqtt5_client_flow
 static int s_do_iot_core_publish_tps_test(struct aws_allocator *allocator, bool use_iot_core_limits) {
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* send pubacks */
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
@@ -2656,7 +2647,7 @@ static int s_do_iot_core_publish_tps_test(struct aws_allocator *allocator, bool 
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     /* send a bunch of tiny publishes */
     for (size_t i = 0; i < IOT_CORE_PUBLISH_TPS_PACKETS; ++i) {
@@ -2686,7 +2677,7 @@ static int s_do_iot_core_publish_tps_test(struct aws_allocator *allocator, bool 
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
 
@@ -2753,7 +2744,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_honor_session_after_success(
         connack_view.session_present = !connect_packet->clean_start;
     }
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 static int s_aws_mqtt5_mock_server_handle_connect_honor_session_unconditional(
@@ -2771,7 +2762,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_honor_session_unconditional(
     connack_view.reason_code = AWS_MQTT5_CRC_SUCCESS;
     connack_view.session_present = !connect_packet->clean_start;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 struct aws_mqtt5_wait_for_n_lifecycle_events_context {
@@ -2836,7 +2827,7 @@ static int s_aws_mqtt5_client_test_init_resume_session_connect_storage(
 
     struct aws_mqtt5_packet_connect_view connect_view = {
         .keep_alive_interval_seconds = 30,
-        .client_id = aws_byte_cursor_from_string(s_client_id),
+        .client_id = aws_byte_cursor_from_string(g_default_client_id),
         .clean_start = false,
     };
 
@@ -2851,7 +2842,7 @@ static int s_do_mqtt5_client_session_resumption_test(
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.client_options.session_behavior = session_behavior;
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
@@ -2899,7 +2890,7 @@ static int s_do_mqtt5_client_session_resumption_test(
     }
 
     ASSERT_SUCCESS(
-        s_verify_received_packet_sequence(&test_context, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
+        aws_verify_received_packet_sequence(&test_context, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
 
     aws_mqtt5_packet_connect_storage_clean_up(&clean_start_connect_storage);
     aws_mqtt5_packet_connect_storage_clean_up(&resume_session_connect_storage);
@@ -3150,7 +3141,7 @@ static int s_aws_mqtt5_server_send_unsuback_on_unsubscribe(
         .reason_codes = s_unsuback_reason_codes,
     };
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_UNSUBACK, &unsuback_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_UNSUBACK, &unsuback_view);
 }
 
 #define FORWARDED_PUBLISH_PACKET_ID 32768
@@ -3171,7 +3162,7 @@ static int s_aws_mqtt5_server_send_puback_and_forward_on_publish(
             .reason_code = AWS_MQTT5_PARC_SUCCESS,
         };
 
-        if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view)) {
+        if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view)) {
             return AWS_OP_ERR;
         }
     }
@@ -3182,7 +3173,7 @@ static int s_aws_mqtt5_server_send_puback_and_forward_on_publish(
         reflect_publish_view.packet_id = FORWARDED_PUBLISH_PACKET_ID;
     }
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &reflect_publish_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &reflect_publish_view);
 }
 
 void s_sub_pub_unsub_publish_complete_fn(
@@ -3273,7 +3264,7 @@ static int s_do_sub_pub_unsub_test(struct aws_allocator *allocator, enum aws_mqt
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     struct aws_mqtt5_client_mock_test_fixture test_context;
     struct aws_mqtt5_sub_pub_unsub_context full_test_context = {
@@ -3301,7 +3292,7 @@ static int s_do_sub_pub_unsub_test(struct aws_allocator *allocator, enum aws_mqt
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_subscribe_storage expected_subscribe_storage;
     AWS_ZERO_STRUCT(expected_subscribe_storage);
@@ -3320,7 +3311,7 @@ static int s_do_sub_pub_unsub_test(struct aws_allocator *allocator, enum aws_mqt
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* verify packets that server received: connect,subscribe, publish, puback(if qos1), unsubscribe */
     struct aws_array_list expected_packets;
@@ -3358,7 +3349,7 @@ static int s_do_sub_pub_unsub_test(struct aws_allocator *allocator, enum aws_mqt
     };
     aws_array_list_push_back(&expected_packets, &unsubscribe_record);
 
-    ASSERT_SUCCESS(s_verify_received_packet_sequence(
+    ASSERT_SUCCESS(aws_verify_received_packet_sequence(
         &test_context, expected_packets.data, aws_array_list_length(&expected_packets)));
 
     /* verify client received the publish that we sent */
@@ -3429,7 +3420,7 @@ static int s_aws_mqtt5_server_send_not_subscribe_unsuback_on_unsubscribe(
         .reason_codes = s_unsubscribe_success_reason_codes,
     };
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_UNSUBACK, &unsuback_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_UNSUBACK, &unsuback_view);
 }
 
 static int s_mqtt5_client_unsubscribe_success_fn(struct aws_allocator *allocator, void *ctx) {
@@ -3438,7 +3429,7 @@ static int s_mqtt5_client_unsubscribe_success_fn(struct aws_allocator *allocator
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_UNSUBSCRIBE] =
         s_aws_mqtt5_server_send_not_subscribe_unsuback_on_unsubscribe;
@@ -3459,7 +3450,7 @@ static int s_mqtt5_client_unsubscribe_success_fn(struct aws_allocator *allocator
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_unsubscribe_view unsubscribe_view = {
         .topic_filters = s_sub_pub_unsub_topic_filters,
@@ -3476,7 +3467,7 @@ static int s_mqtt5_client_unsubscribe_success_fn(struct aws_allocator *allocator
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -3537,14 +3528,14 @@ static void s_aws_mqtt5_mock_server_send_qos1_publish(
         .packet_id = s_puback_packet_id,
     };
 
-    s_aws_mqtt5_mock_server_send_packet(mock_server, AWS_MQTT5_PT_PUBLISH, &qos1_publish_view);
+    aws_mqtt5_mock_server_send_packet(mock_server, AWS_MQTT5_PT_PUBLISH, &qos1_publish_view);
 }
 
 static int s_aws_mqtt5_server_send_qos1_publish_on_connect(
     void *packet,
     struct aws_mqtt5_server_mock_connection_context *connection,
     void *user_data) {
-    int result = s_aws_mqtt5_mock_server_handle_connect_always_succeed(packet, connection, user_data);
+    int result = aws_mqtt5_mock_server_handle_connect_always_succeed(packet, connection, user_data);
 
     struct aws_mqtt5_server_send_qos1_publish_context *test_context = user_data;
     test_context->connack_sent = true;
@@ -3571,7 +3562,7 @@ static int mqtt5_client_receive_qos1_return_puback_test_fn(struct aws_allocator 
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* mock server sends a PUBLISH packet to the client */
     test_options.server_function_table.service_task_fn = s_aws_mqtt5_mock_server_send_qos1_publish;
@@ -3599,13 +3590,13 @@ static int mqtt5_client_receive_qos1_return_puback_test_fn(struct aws_allocator 
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     s_publish_qos1_wait_for_puback(&publish_context);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -3628,7 +3619,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_session_present(
     connack_view.reason_code = AWS_MQTT5_CRC_SUCCESS;
     connack_view.session_present = true;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 /* When client receives a CONNACK with existing session state when one isn't present it should disconnect */
@@ -3638,7 +3629,7 @@ static int mqtt5_client_receive_nonexisting_session_state_fn(struct aws_allocato
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* mock server returns a CONNACK indicating a session is being resumed */
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
@@ -3659,14 +3650,14 @@ static int mqtt5_client_receive_nonexisting_session_state_fn(struct aws_allocato
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     enum aws_mqtt5_client_state expected_states[] = {
         AWS_MCS_CONNECTING,
         AWS_MCS_MQTT_CONNECT,
         AWS_MCS_CHANNEL_SHUTDOWN,
     };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -3707,7 +3698,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_assigned_client_id(
         test_context->assigned_client_id_checked = true;
     }
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 /*
@@ -3720,7 +3711,7 @@ static int mqtt5_client_receive_assigned_client_id_fn(struct aws_allocator *allo
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* Empty the Client ID for connect */
     test_options.connect_options.client_id.len = 0;
@@ -3747,7 +3738,7 @@ static int mqtt5_client_receive_assigned_client_id_fn(struct aws_allocator *allo
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_byte_cursor assigned_client_id = aws_byte_cursor_from_c_str(s_receive_assigned_client_id_client_id);
     struct aws_byte_cursor negotiated_settings_client_id =
@@ -3761,18 +3752,18 @@ static int mqtt5_client_receive_assigned_client_id_fn(struct aws_allocator *allo
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* Check for Assigned Client ID on reconnect */
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     ASSERT_TRUE(assinged_id_context.assigned_client_id_checked);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -3802,7 +3793,7 @@ static int s_aws_mqtt5_mock_server_handle_publish_no_puback_on_first_connect(
         struct aws_mqtt5_packet_puback_view puback_view = {
             .packet_id = publish_view->packet_id,
         };
-        return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view);
+        return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view);
     }
 
     return AWS_OP_SUCCESS;
@@ -3854,7 +3845,7 @@ static int mqtt5_client_no_session_after_client_stop_fn(struct aws_allocator *al
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* Set to rejoin */
     test_options.client_options.session_behavior = AWS_MQTT5_CSBT_REJOIN_POST_SUCCESS;
@@ -3879,7 +3870,7 @@ static int mqtt5_client_no_session_after_client_stop_fn(struct aws_allocator *al
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     for (size_t i = 0; i < TEST_PUBLISH_COUNT; ++i) {
         struct aws_mqtt5_packet_publish_view qos1_publish_view = {
@@ -3908,11 +3899,11 @@ static int mqtt5_client_no_session_after_client_stop_fn(struct aws_allocator *al
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     aws_mutex_lock(&test_context.lock);
     size_t event_count = aws_array_list_length(&test_context.lifecycle_events);
@@ -3923,7 +3914,7 @@ static int mqtt5_client_no_session_after_client_stop_fn(struct aws_allocator *al
     ASSERT_FALSE(record->connack_storage.storage_view.session_present);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -3939,7 +3930,7 @@ static int mqtt5_client_restore_session_on_ping_timeout_reconnect_fn(struct aws_
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* Set to rejoin */
     test_options.client_options.session_behavior = AWS_MQTT5_CSBT_REJOIN_POST_SUCCESS;
@@ -3969,7 +3960,7 @@ static int mqtt5_client_restore_session_on_ping_timeout_reconnect_fn(struct aws_
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     for (size_t i = 0; i < TEST_PUBLISH_COUNT; ++i) {
         struct aws_mqtt5_packet_publish_view qos1_publish_view = {
@@ -4000,12 +3991,12 @@ static int mqtt5_client_restore_session_on_ping_timeout_reconnect_fn(struct aws_
     s_wait_for_disconnection_lifecycle_event(&test_context);
 
     /* Reconnect from a disconnect automatically */
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     s_wait_for_n_successful_publishes(&wait_context);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     enum aws_mqtt5_client_state expected_states[] = {
         AWS_MCS_CONNECTING,
@@ -4021,7 +4012,7 @@ static int mqtt5_client_restore_session_on_ping_timeout_reconnect_fn(struct aws_
         AWS_MCS_STOPPED,
     };
 
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
+    ASSERT_SUCCESS(aws_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -4040,7 +4031,7 @@ static int mqtt5_client_discard_session_on_server_clean_start_fn(struct aws_allo
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     /* Set to rejoin */
     test_options.client_options.session_behavior = AWS_MQTT5_CSBT_REJOIN_POST_SUCCESS;
@@ -4062,7 +4053,7 @@ static int mqtt5_client_discard_session_on_server_clean_start_fn(struct aws_allo
 
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     for (size_t i = 0; i < TEST_PUBLISH_COUNT; ++i) {
         struct aws_mqtt5_packet_publish_view qos1_publish_view = {
@@ -4091,17 +4082,17 @@ static int mqtt5_client_discard_session_on_server_clean_start_fn(struct aws_allo
 
     /* Disconnect with unacked publishes */
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* Reconnect with a Client Stored Session */
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     /* Provide time for Client to process any queued operations */
     aws_thread_current_sleep(1000000000);
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     /* Check that no publishes were resent after the initial batch on first connect */
     ASSERT_INT_EQUALS(test_context.publishes_received, TEST_PUBLISH_COUNT);
@@ -4179,7 +4170,7 @@ static int s_mqtt5_client_statistics_subscribe_fn(struct aws_allocator *allocato
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_SUBSCRIBE] =
         s_aws_mqtt5_server_send_suback_on_subscribe;
@@ -4202,7 +4193,7 @@ static int s_mqtt5_client_statistics_subscribe_fn(struct aws_allocator *allocato
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_subscribe_view subscribe_view = {
         .subscriptions = s_subscriptions,
@@ -4219,7 +4210,7 @@ static int s_mqtt5_client_statistics_subscribe_fn(struct aws_allocator *allocato
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_verify_client_statistics(
         &test_context, s_subscribe_test_statistics, AWS_ARRAY_SIZE(s_subscribe_test_statistics)));
@@ -4253,7 +4244,7 @@ static int s_mqtt5_client_statistics_unsubscribe_fn(struct aws_allocator *alloca
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_UNSUBSCRIBE] =
         s_aws_mqtt5_server_send_not_subscribe_unsuback_on_unsubscribe;
@@ -4274,7 +4265,7 @@ static int s_mqtt5_client_statistics_unsubscribe_fn(struct aws_allocator *alloca
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_unsubscribe_view unsubscribe_view = {
         .topic_filters = s_sub_pub_unsub_topic_filters,
@@ -4291,7 +4282,7 @@ static int s_mqtt5_client_statistics_unsubscribe_fn(struct aws_allocator *alloca
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_verify_client_statistics(
         &test_context, s_unsubscribe_test_statistics, AWS_ARRAY_SIZE(s_unsubscribe_test_statistics)));
@@ -4327,7 +4318,7 @@ static int s_do_mqtt5_client_statistics_publish_test(
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
         s_aws_mqtt5_server_send_puback_and_forward_on_publish;
@@ -4348,7 +4339,7 @@ static int s_do_mqtt5_client_statistics_publish_test(
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_publish_view publish_view = {
         .qos = qos,
@@ -4374,7 +4365,7 @@ static int s_do_mqtt5_client_statistics_publish_test(
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_verify_client_statistics(&test_context, expected_stats, expected_stats_count));
 
@@ -4469,7 +4460,7 @@ static int s_aws_mqtt5_server_disconnect_on_first_publish_puback_after(
             .reason_code = AWS_MQTT5_PARC_SUCCESS,
         };
 
-        if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view)) {
+        if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBACK, &puback_view)) {
             return AWS_OP_ERR;
         }
     }
@@ -4483,7 +4474,7 @@ static int s_mqtt5_client_statistics_publish_qos1_requeue_fn(struct aws_allocato
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.client_options.session_behavior = AWS_MQTT5_CSBT_REJOIN_POST_SUCCESS;
 
@@ -4508,7 +4499,7 @@ static int s_mqtt5_client_statistics_publish_qos1_requeue_fn(struct aws_allocato
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_publish_view publish_view = {
         .qos = AWS_MQTT5_QOS_AT_LEAST_ONCE,
@@ -4534,7 +4525,7 @@ static int s_mqtt5_client_statistics_publish_qos1_requeue_fn(struct aws_allocato
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_verify_client_statistics(
         &test_context, s_publish_qos1_requeue_test_statistics, AWS_ARRAY_SIZE(s_publish_qos1_requeue_test_statistics)));
@@ -4573,7 +4564,7 @@ static int s_aws_mqtt5_server_send_multiple_publishes(
                 },
         };
 
-        if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
+        if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
             return AWS_OP_ERR;
         }
     }
@@ -4636,7 +4627,7 @@ static int s_mqtt5_client_puback_ordering_fn(struct aws_allocator *allocator, vo
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
         s_aws_mqtt5_server_send_multiple_publishes;
@@ -4657,7 +4648,7 @@ static int s_mqtt5_client_puback_ordering_fn(struct aws_allocator *allocator, vo
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_publish_view publish_view = {
         .qos = AWS_MQTT5_QOS_AT_MOST_ONCE,
@@ -4679,7 +4670,7 @@ static int s_mqtt5_client_puback_ordering_fn(struct aws_allocator *allocator, vo
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_verify_mock_puback_order(&test_context));
 
@@ -4737,7 +4728,7 @@ static int s_aws_mqtt5_mock_server_reflect_publish(
         reflected_view.packet_id = 1;
     }
 
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &reflected_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &reflected_view)) {
         return AWS_OP_ERR;
     }
 
@@ -4812,7 +4803,7 @@ static int s_mqtt5_client_listeners_fn(struct aws_allocator *allocator, void *ct
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] = s_aws_mqtt5_mock_server_reflect_publish;
 
@@ -4837,7 +4828,7 @@ static int s_mqtt5_client_listeners_fn(struct aws_allocator *allocator, void *ct
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_publish_view qos0_publish_view = {
         .qos = AWS_MQTT5_QOS_AT_MOST_ONCE,
@@ -4916,7 +4907,7 @@ static int s_mqtt5_client_listeners_fn(struct aws_allocator *allocator, void *ct
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     s_aws_mqtt5_listeners_test_context_clean_up(&full_test_context);
@@ -5155,7 +5146,7 @@ static int s_mqtt5_client_offline_operation_submission_fail_all_fn(struct aws_al
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.client_options.offline_queue_behavior = AWS_MQTT5_COQBT_FAIL_ALL_ON_DISCONNECT;
 
@@ -5196,7 +5187,7 @@ static int s_mqtt5_client_offline_operation_submission_fail_qos0_fn(struct aws_a
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.client_options.offline_queue_behavior = AWS_MQTT5_COQBT_FAIL_QOS0_PUBLISH_ON_DISCONNECT;
 
@@ -5242,7 +5233,7 @@ static int s_mqtt5_client_offline_operation_submission_fail_non_qos1_fn(struct a
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.client_options.offline_queue_behavior = AWS_MQTT5_COQBT_FAIL_NON_QOS1_PUBLISH_ON_DISCONNECT;
 
@@ -5286,7 +5277,7 @@ static int s_mqtt5_client_offline_operation_submission_then_connect_fn(struct aw
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_PUBLISH] =
         s_aws_mqtt5_server_send_puback_and_forward_on_publish;
@@ -5319,7 +5310,7 @@ static int s_mqtt5_client_offline_operation_submission_then_connect_fn(struct aw
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
 
@@ -5399,7 +5390,7 @@ static int s_aws_mqtt5_server_send_aliased_publish_sequence(
         .packet_id = subscribe_view->packet_id, .reason_code_count = 1, .reason_codes = s_alias_reason_codes};
 
     // just to be thorough, send a suback
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_SUBACK, &suback_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_SUBACK, &suback_view)) {
         return AWS_OP_ERR;
     }
 
@@ -5416,13 +5407,13 @@ static int s_aws_mqtt5_server_send_aliased_publish_sequence(
     };
 
     // establish an alias with id 1
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
         return AWS_OP_ERR;
     }
 
     // alias alone
     AWS_ZERO_STRUCT(publish_view.topic);
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
         return AWS_OP_ERR;
     }
 
@@ -5430,13 +5421,13 @@ static int s_aws_mqtt5_server_send_aliased_publish_sequence(
     publish_view.topic.ptr = s_alias_topic2;
     publish_view.topic.len = AWS_ARRAY_SIZE(s_alias_topic2) - 1;
 
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
         return AWS_OP_ERR;
     }
 
     // alias alone
     AWS_ZERO_STRUCT(publish_view.topic);
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
         return AWS_OP_ERR;
     }
 
@@ -5510,7 +5501,7 @@ static int s_mqtt5_client_inbound_alias_success_fn(struct aws_allocator *allocat
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_SUBSCRIBE] =
         s_aws_mqtt5_server_send_aliased_publish_sequence;
@@ -5541,7 +5532,7 @@ static int s_mqtt5_client_inbound_alias_success_fn(struct aws_allocator *allocat
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_subscribe_view subscribe = {
         .subscriptions = s_alias_subscriptions,
@@ -5554,7 +5545,7 @@ static int s_mqtt5_client_inbound_alias_success_fn(struct aws_allocator *allocat
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_verify_aliased_publish_sequence(&full_test_context));
 
@@ -5594,7 +5585,7 @@ static int s_aws_mqtt5_server_send_aliased_publish_failure(
         .packet_id = subscribe_view->packet_id, .reason_code_count = 1, .reason_codes = s_alias_reason_codes};
 
     // just to be thorough, send a suback
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_SUBACK, &suback_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_SUBACK, &suback_view)) {
         return AWS_OP_ERR;
     }
 
@@ -5622,7 +5613,7 @@ static int s_aws_mqtt5_server_send_aliased_publish_failure(
         .packet_id = 1, .qos = AWS_MQTT5_QOS_AT_LEAST_ONCE, .topic = topic_cursor, .topic_alias = &alias_id};
 
     // establish an alias with id 1
-    if (s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
+    if (aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_PUBLISH, &publish_view)) {
         return AWS_OP_ERR;
     }
 
@@ -5662,7 +5653,7 @@ static int s_do_inbound_alias_failure_test(
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_SUBSCRIBE] =
         s_aws_mqtt5_server_send_aliased_publish_failure;
@@ -5693,7 +5684,7 @@ static int s_do_inbound_alias_failure_test(
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     struct aws_mqtt5_packet_subscribe_view subscribe = {
         .subscriptions = s_alias_subscriptions,
@@ -5706,7 +5697,7 @@ static int s_do_inbound_alias_failure_test(
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -5788,7 +5779,7 @@ static int s_aws_mqtt5_mock_server_handle_connect_allow_aliasing(
     uint16_t topic_alias_maximum = SEQUENCE_TEST_CACHE_SIZE;
     connack_view.topic_alias_maximum = &topic_alias_maximum;
 
-    return s_aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
+    return aws_mqtt5_mock_server_send_packet(connection, AWS_MQTT5_PT_CONNACK, &connack_view);
 }
 
 static int s_do_mqtt5_client_outbound_alias_failure_test(
@@ -5798,7 +5789,7 @@ static int s_do_mqtt5_client_outbound_alias_failure_test(
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
         s_aws_mqtt5_mock_server_handle_connect_allow_aliasing;
@@ -5820,7 +5811,7 @@ static int s_do_mqtt5_client_outbound_alias_failure_test(
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     uint16_t topic_alias = 1;
     struct aws_mqtt5_packet_publish_view packet_publish_view = {
@@ -5849,7 +5840,7 @@ static int s_do_mqtt5_client_outbound_alias_failure_test(
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -5976,7 +5967,7 @@ static int s_perform_outbound_alias_sequence_test(
     aws_mqtt_library_init(allocator);
 
     struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
+    aws_mqtt5_client_test_init_default_options(&test_options);
 
     test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
         s_aws_mqtt5_mock_server_handle_connect_allow_aliasing;
@@ -5998,7 +5989,7 @@ static int s_perform_outbound_alias_sequence_test(
     struct aws_mqtt5_client *client = test_context.client;
     ASSERT_SUCCESS(aws_mqtt5_client_start(client));
 
-    s_wait_for_connected_lifecycle_event(&test_context);
+    aws_wait_for_connected_lifecycle_event(&test_context);
 
     ASSERT_SUCCESS(s_perform_outbound_alias_sequence(&test_context, publishes, publish_count));
 
@@ -6038,7 +6029,7 @@ static int s_perform_outbound_alias_sequence_test(
 
     ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
 
-    s_wait_for_stopped_lifecycle_event(&test_context);
+    aws_wait_for_stopped_lifecycle_event(&test_context);
 
     aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
     aws_mqtt_library_clean_up();
@@ -6091,909 +6082,3 @@ static int s_mqtt5_client_outbound_alias_lru_success_a_b_c_br_cr_a_fn(struct aws
 AWS_TEST_CASE(
     mqtt5_client_outbound_alias_lru_success_a_b_c_br_cr_a,
     s_mqtt5_client_outbound_alias_lru_success_a_b_c_br_cr_a_fn)
-
-enum aws_mqtt3_lifecycle_event_type {
-    AWS_MQTT3_LET_CONNECTION_COMPLETE,
-    AWS_MQTT3_LET_INTERRUPTED,
-    AWS_MQTT3_LET_RESUMED,
-    AWS_MQTT3_LET_CLOSED,
-    AWS_MQTT3_LET_DISCONNECTION_COMPLETE,
-};
-
-struct aws_mqtt3_lifecycle_event {
-    enum aws_mqtt3_lifecycle_event_type type;
-
-    uint64_t timestamp;
-    int error_code;
-    enum aws_mqtt_connect_return_code return_code;
-    bool session_present;
-};
-
-struct aws_mqtt3_to_mqtt5_adapter_test_fixture_config {
-    aws_mqtt_client_on_connection_interrupted_fn *on_interrupted;
-    aws_mqtt_client_on_connection_resumed_fn *on_resumed;
-    aws_mqtt_client_on_connection_closed_fn *on_closed;
-
-    void *callback_user_data;
-};
-
-struct aws_mqtt3_to_mqtt5_adapter_test_fixture {
-    struct aws_mqtt5_client_mock_test_fixture mqtt5_fixture;
-
-    struct aws_mqtt_client_connection *connection;
-
-    struct aws_array_list lifecycle_events;
-
-    struct aws_mutex lock;
-    struct aws_condition_variable signal;
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture_config config;
-};
-
-static void s_init_adapter_connection_options_from_fixture(
-    struct aws_mqtt_connection_options *connection_options,
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture) {
-    AWS_ZERO_STRUCT(*connection_options);
-
-    connection_options->host_name = aws_byte_cursor_from_c_str(fixture->mqtt5_fixture.endpoint.address);
-    connection_options->port = fixture->mqtt5_fixture.endpoint.port;
-    connection_options->socket_options = &fixture->mqtt5_fixture.socket_options;
-    connection_options->keep_alive_time_secs = 30;
-    connection_options->ping_timeout_ms = 10000;
-    connection_options->clean_session = true;
-}
-
-struct n_lifeycle_event_wait_context {
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture;
-    enum aws_mqtt3_lifecycle_event_type type;
-    size_t count;
-};
-
-static bool s_wait_for_n_adapter_lifecycle_events_predicate(void *context) {
-    struct n_lifeycle_event_wait_context *wait_context = context;
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture = wait_context->fixture;
-
-    size_t actual_count = 0;
-    size_t event_count = aws_array_list_length(&fixture->lifecycle_events);
-    for (size_t i = 0; i < event_count; ++i) {
-        struct aws_mqtt3_lifecycle_event *actual_event = NULL;
-        aws_array_list_get_at_ptr(&fixture->lifecycle_events, (void **)(&actual_event), i);
-        if (actual_event->type == wait_context->type) {
-            ++actual_count;
-        }
-    }
-
-    return actual_count >= wait_context->count;
-}
-
-static void s_wait_for_n_adapter_lifecycle_events(
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture,
-    enum aws_mqtt3_lifecycle_event_type type,
-    size_t count) {
-    struct n_lifeycle_event_wait_context wait_context = {
-        .fixture = fixture,
-        .type = type,
-        .count = count,
-    };
-
-    aws_mutex_lock(&fixture->lock);
-    aws_condition_variable_wait_pred(
-        &fixture->signal, &fixture->lock, s_wait_for_n_adapter_lifecycle_events_predicate, &wait_context);
-    aws_mutex_unlock(&fixture->lock);
-}
-
-static int s_aws_mqtt3_to_mqtt5_adapter_test_fixture_verify_lifecycle_sequence(
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture,
-    size_t event_count,
-    struct aws_mqtt3_lifecycle_event *expected_events) {
-
-    aws_mutex_lock(&fixture->lock);
-
-    size_t actual_event_count = aws_array_list_length(&fixture->lifecycle_events);
-    ASSERT_TRUE(event_count <= actual_event_count);
-
-    for (size_t i = 0; i < event_count; ++i) {
-        struct aws_mqtt3_lifecycle_event *expected_event = expected_events + i;
-        struct aws_mqtt3_lifecycle_event *actual_event = NULL;
-        aws_array_list_get_at_ptr(&fixture->lifecycle_events, (void **)(&actual_event), i);
-
-        ASSERT_INT_EQUALS(actual_event->type, expected_event->type);
-        ASSERT_INT_EQUALS(actual_event->error_code, expected_event->error_code);
-        ASSERT_INT_EQUALS(actual_event->return_code, expected_event->return_code);
-        ASSERT_TRUE(actual_event->session_present == expected_event->session_present);
-    }
-
-    aws_mutex_unlock(&fixture->lock);
-
-    return AWS_OP_SUCCESS;
-}
-
-static void s_aws_mqtt3_to_mqtt5_adapter_test_fixture_closed_handler(
-    struct aws_mqtt_client_connection *connection,
-    struct on_connection_closed_data *data,
-    void *userdata) {
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture = userdata;
-
-    /* record the event */
-    struct aws_mqtt3_lifecycle_event event;
-    AWS_ZERO_STRUCT(event);
-
-    event.type = AWS_MQTT3_LET_CLOSED;
-    aws_high_res_clock_get_ticks(&event.timestamp);
-
-    aws_mutex_lock(&fixture->lock);
-    aws_array_list_push_back(&fixture->lifecycle_events, &event);
-    aws_mutex_unlock(&fixture->lock);
-    aws_condition_variable_notify_all(&fixture->signal);
-
-    /* invoke user callback if registered */
-    if (fixture->config.on_closed) {
-        (*fixture->config.on_closed)(connection, data, fixture->config.callback_user_data);
-    }
-}
-
-static void s_aws_mqtt3_to_mqtt5_adapter_test_fixture_interrupted_handler(
-    struct aws_mqtt_client_connection *connection,
-    int error_code,
-    void *userdata) {
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture = userdata;
-
-    /* record the event */
-    struct aws_mqtt3_lifecycle_event event;
-    AWS_ZERO_STRUCT(event);
-
-    event.type = AWS_MQTT3_LET_INTERRUPTED;
-    aws_high_res_clock_get_ticks(&event.timestamp);
-    event.error_code = error_code;
-
-    aws_mutex_lock(&fixture->lock);
-    aws_array_list_push_back(&fixture->lifecycle_events, &event);
-    aws_mutex_unlock(&fixture->lock);
-    aws_condition_variable_notify_all(&fixture->signal);
-
-    /* invoke user callback if registered */
-    if (fixture->config.on_interrupted) {
-        (*fixture->config.on_interrupted)(connection, error_code, fixture->config.callback_user_data);
-    }
-}
-
-static void s_aws_mqtt3_to_mqtt5_adapter_test_fixture_resumed_handler(
-    struct aws_mqtt_client_connection *connection,
-    enum aws_mqtt_connect_return_code return_code,
-    bool session_present,
-    void *userdata) {
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture = userdata;
-
-    /* record the event */
-    struct aws_mqtt3_lifecycle_event event;
-    AWS_ZERO_STRUCT(event);
-
-    event.type = AWS_MQTT3_LET_RESUMED;
-    aws_high_res_clock_get_ticks(&event.timestamp);
-    event.return_code = return_code;
-    event.session_present = session_present;
-
-    aws_mutex_lock(&fixture->lock);
-    aws_array_list_push_back(&fixture->lifecycle_events, &event);
-    aws_mutex_unlock(&fixture->lock);
-    aws_condition_variable_notify_all(&fixture->signal);
-
-    /* invoke user callback if registered */
-    if (fixture->config.on_resumed) {
-        (*fixture->config.on_resumed)(connection, return_code, session_present, fixture->config.callback_user_data);
-    }
-}
-
-static void s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_connection_complete(
-    struct aws_mqtt_client_connection *connection,
-    int error_code,
-    enum aws_mqtt_connect_return_code return_code,
-    bool session_present,
-    void *user_data) {
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture = user_data;
-
-    struct aws_mqtt3_lifecycle_event event;
-    AWS_ZERO_STRUCT(event);
-
-    event.type = AWS_MQTT3_LET_CONNECTION_COMPLETE;
-    aws_high_res_clock_get_ticks(&event.timestamp);
-    event.error_code = error_code;
-    event.return_code = return_code;
-    event.session_present = session_present;
-
-    aws_mutex_lock(&fixture->lock);
-    aws_array_list_push_back(&fixture->lifecycle_events, &event);
-    aws_mutex_unlock(&fixture->lock);
-    aws_condition_variable_notify_all(&fixture->signal);
-}
-
-static void s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_disconnection_complete(
-    struct aws_mqtt_client_connection *connection,
-    void *user_data) {
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture = user_data;
-
-    struct aws_mqtt3_lifecycle_event event;
-    AWS_ZERO_STRUCT(event);
-
-    event.type = AWS_MQTT3_LET_DISCONNECTION_COMPLETE;
-    aws_high_res_clock_get_ticks(&event.timestamp);
-
-    aws_mutex_lock(&fixture->lock);
-    aws_array_list_push_back(&fixture->lifecycle_events, &event);
-    aws_mutex_unlock(&fixture->lock);
-    aws_condition_variable_notify_all(&fixture->signal);
-}
-
-int aws_mqtt3_to_mqtt5_adapter_test_fixture_init(
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture,
-    struct aws_allocator *allocator,
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options *mqtt5_fixture_config,
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture_config *config) {
-    AWS_ZERO_STRUCT(*fixture);
-
-    if (aws_mqtt5_client_mock_test_fixture_init(&fixture->mqtt5_fixture, allocator, mqtt5_fixture_config)) {
-        return AWS_OP_ERR;
-    }
-
-    fixture->connection = aws_mqtt_client_connection_new_from_mqtt5_client(fixture->mqtt5_fixture.client);
-    if (fixture->connection == NULL) {
-        return AWS_OP_ERR;
-    }
-
-    aws_array_list_init_dynamic(&fixture->lifecycle_events, allocator, 10, sizeof(struct aws_mqtt3_lifecycle_event));
-
-    aws_mutex_init(&fixture->lock);
-    aws_condition_variable_init(&fixture->signal);
-
-    if (config) {
-        fixture->config = *config;
-    }
-
-    aws_mqtt_client_connection_set_connection_closed_handler(
-        fixture->connection, s_aws_mqtt3_to_mqtt5_adapter_test_fixture_closed_handler, fixture);
-    aws_mqtt_client_connection_set_connection_interruption_handlers(
-        fixture->connection,
-        s_aws_mqtt3_to_mqtt5_adapter_test_fixture_interrupted_handler,
-        fixture,
-        s_aws_mqtt3_to_mqtt5_adapter_test_fixture_resumed_handler,
-        fixture);
-
-    return AWS_OP_SUCCESS;
-}
-
-void aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(struct aws_mqtt3_to_mqtt5_adapter_test_fixture *fixture) {
-    aws_mqtt_client_connection_release(fixture->connection);
-
-    aws_mqtt5_client_mock_test_fixture_clean_up(&fixture->mqtt5_fixture);
-
-    aws_array_list_clean_up(&fixture->lifecycle_events);
-
-    aws_mutex_clean_up(&fixture->lock);
-    aws_condition_variable_clean_up(&fixture->signal);
-}
-
-void s_mqtt3to5_lifecycle_event_callback(const struct aws_mqtt5_client_lifecycle_event *event) {
-    (void)event;
-}
-
-void s_mqtt3to5_publish_received_callback(const struct aws_mqtt5_packet_publish_view *publish, void *user_data) {
-    (void)publish;
-    (void)user_data;
-}
-
-static int s_do_mqtt3to5_adapter_create_destroy(struct aws_allocator *allocator, uint64_t sleep_nanos) {
-    aws_mqtt_library_init(allocator);
-
-    struct aws_mqtt5_packet_connect_view local_connect_options = {
-        .keep_alive_interval_seconds = 30,
-        .clean_start = true,
-    };
-
-    struct aws_mqtt5_client_options client_options = {
-        .connect_options = &local_connect_options,
-        .lifecycle_event_handler = s_mqtt3to5_lifecycle_event_callback,
-        .lifecycle_event_handler_user_data = NULL,
-        .publish_received_handler = s_mqtt3to5_publish_received_callback,
-        .publish_received_handler_user_data = NULL,
-        .ping_timeout_ms = 10000,
-    };
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_config = {
-        .client_options = &client_options,
-    };
-
-    struct aws_mqtt5_client_mock_test_fixture test_fixture;
-    AWS_ZERO_STRUCT(test_fixture);
-
-    ASSERT_SUCCESS(aws_mqtt5_client_mock_test_fixture_init(&test_fixture, allocator, &test_fixture_config));
-
-    struct aws_mqtt_client_connection *connection =
-        aws_mqtt_client_connection_new_from_mqtt5_client(test_fixture.client);
-
-    if (sleep_nanos > 0) {
-        /* sleep a little just to let the listener attachment resolve */
-        aws_thread_current_sleep(aws_timestamp_convert(1, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL));
-    }
-
-    aws_mqtt_client_connection_release(connection);
-
-    if (sleep_nanos > 0) {
-        /* sleep a little just to let the listener detachment resolve */
-        aws_thread_current_sleep(aws_timestamp_convert(1, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL));
-    }
-
-    aws_mqtt5_client_mock_test_fixture_clean_up(&test_fixture);
-
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_mqtt3to5_adapter_create_destroy_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    ASSERT_SUCCESS(s_do_mqtt3to5_adapter_create_destroy(allocator, 0));
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_create_destroy, s_mqtt3to5_adapter_create_destroy_fn)
-
-static int s_mqtt3to5_adapter_create_destroy_delayed_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    ASSERT_SUCCESS(s_do_mqtt3to5_adapter_create_destroy(
-        allocator, aws_timestamp_convert(1, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL)));
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_create_destroy_delayed, s_mqtt3to5_adapter_create_destroy_delayed_fn)
-
-typedef int (*mqtt3to5_adapter_config_test_setup_fn)(
-    struct aws_allocator *allocator,
-    struct aws_mqtt_client_connection *adapter,
-    struct aws_mqtt5_packet_connect_storage *expected_connect);
-
-static int s_do_mqtt3to5_adapter_config_test(
-    struct aws_allocator *allocator,
-    mqtt3to5_adapter_config_test_setup_fn setup_fn) {
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture test_fixture;
-    ASSERT_SUCCESS(aws_mqtt3_to_mqtt5_adapter_test_fixture_init(&test_fixture, allocator, &test_fixture_options, NULL));
-
-    struct aws_mqtt5_client *client = test_fixture.mqtt5_fixture.client;
-
-    struct aws_mqtt_client_connection *adapter = test_fixture.connection;
-
-    struct aws_mqtt5_packet_connect_storage expected_connect_storage;
-    ASSERT_SUCCESS((*setup_fn)(allocator, adapter, &expected_connect_storage));
-
-    ASSERT_SUCCESS(aws_mqtt5_client_start(client));
-
-    s_wait_for_connected_lifecycle_event(&test_fixture.mqtt5_fixture);
-
-    ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
-
-    s_wait_for_stopped_lifecycle_event(&test_fixture.mqtt5_fixture);
-
-    struct aws_mqtt5_mock_server_packet_record expected_packets[] = {
-        {
-            .packet_type = AWS_MQTT5_PT_CONNECT,
-            .packet_storage = &expected_connect_storage,
-        },
-    };
-    ASSERT_SUCCESS(s_verify_received_packet_sequence(
-        &test_fixture.mqtt5_fixture, expected_packets, AWS_ARRAY_SIZE(expected_packets)));
-
-    aws_mqtt5_packet_connect_storage_clean_up(&expected_connect_storage);
-
-    aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(&test_fixture);
-
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_mqtt3to5_adapter_set_will_setup(
-    struct aws_allocator *allocator,
-    struct aws_mqtt_client_connection *adapter,
-    struct aws_mqtt5_packet_connect_storage *expected_connect) {
-
-    struct aws_byte_cursor topic_cursor = {
-        .ptr = s_alias_topic1,
-        .len = AWS_ARRAY_SIZE(s_alias_topic1) - 1,
-    };
-
-    struct aws_byte_cursor payload_cursor = {
-        .ptr = s_sub_pub_unsub_publish_payload,
-        .len = AWS_ARRAY_SIZE(s_sub_pub_unsub_publish_payload) - 1,
-    };
-
-    ASSERT_SUCCESS(
-        aws_mqtt_client_connection_set_will(adapter, &topic_cursor, AWS_MQTT_QOS_AT_LEAST_ONCE, true, &payload_cursor));
-
-    struct aws_mqtt5_packet_publish_view expected_will = {
-        .payload = payload_cursor,
-        .qos = AWS_MQTT5_QOS_AT_LEAST_ONCE,
-        .retain = true,
-        .topic = topic_cursor,
-    };
-
-    struct aws_mqtt5_packet_connect_view expected_connect_view = {
-        .client_id = aws_byte_cursor_from_string(s_client_id),
-        .keep_alive_interval_seconds = 30,
-        .clean_start = true,
-        .will = &expected_will,
-    };
-
-    ASSERT_SUCCESS(aws_mqtt5_packet_connect_storage_init(expected_connect, allocator, &expected_connect_view));
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_mqtt3to5_adapter_set_will_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    ASSERT_SUCCESS(s_do_mqtt3to5_adapter_config_test(allocator, s_mqtt3to5_adapter_set_will_setup));
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_set_will, s_mqtt3to5_adapter_set_will_fn)
-
-AWS_STATIC_STRING_FROM_LITERAL(s_username, "MyUsername");
-AWS_STATIC_STRING_FROM_LITERAL(s_password, "TopTopSecret");
-
-static int s_mqtt3to5_adapter_set_login_setup(
-    struct aws_allocator *allocator,
-    struct aws_mqtt_client_connection *adapter,
-    struct aws_mqtt5_packet_connect_storage *expected_connect) {
-
-    struct aws_byte_cursor username_cursor = aws_byte_cursor_from_string(s_username);
-    struct aws_byte_cursor password_cursor = aws_byte_cursor_from_string(s_password);
-
-    ASSERT_SUCCESS(aws_mqtt_client_connection_set_login(adapter, &username_cursor, &password_cursor));
-
-    struct aws_mqtt5_packet_connect_view expected_connect_view = {
-        .client_id = aws_byte_cursor_from_string(s_client_id),
-        .keep_alive_interval_seconds = 30,
-        .clean_start = true,
-        .username = &username_cursor,
-        .password = &password_cursor,
-    };
-
-    ASSERT_SUCCESS(aws_mqtt5_packet_connect_storage_init(expected_connect, allocator, &expected_connect_view));
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_mqtt3to5_adapter_set_login_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    ASSERT_SUCCESS(s_do_mqtt3to5_adapter_config_test(allocator, s_mqtt3to5_adapter_set_login_setup));
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_set_login, s_mqtt3to5_adapter_set_login_fn)
-
-static int s_mqtt3to5_adapter_set_reconnect_timeout_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    /*
-     * This is a variant of the mqtt5_client_reconnect_failure_backoff test.
-     *
-     * The primary change is that we configure the mqtt5 client with "wrong" (fast) reconnect delays and then use
-     * the adapter API to configure with the "right" ones that will let the test pass.
-     */
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    /* backoff delay sequence: 500, 1000, 2000, 4000, 5000, ... */
-    test_options.client_options.retry_jitter_mode = AWS_EXPONENTIAL_BACKOFF_JITTER_NONE;
-    test_options.client_options.min_reconnect_delay_ms = 10;
-    test_options.client_options.max_reconnect_delay_ms = 50;
-    test_options.client_options.min_connected_time_to_reset_reconnect_delay_ms = RECONNECT_TEST_BACKOFF_RESET_DELAY;
-
-    test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
-        s_aws_mqtt5_mock_server_handle_connect_always_fail;
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt5_client_mock_test_fixture test_context;
-    ASSERT_SUCCESS(aws_mqtt5_client_mock_test_fixture_init(&test_context, allocator, &test_fixture_options));
-
-    struct aws_mqtt5_client *client = test_context.client;
-
-    struct aws_mqtt_client_connection *adapter = aws_mqtt_client_connection_new_from_mqtt5_client(client);
-
-    aws_mqtt_client_connection_set_reconnect_timeout(adapter, RECONNECT_TEST_MIN_BACKOFF, RECONNECT_TEST_MAX_BACKOFF);
-
-    ASSERT_SUCCESS(aws_mqtt5_client_start(client));
-
-    s_wait_for_n_connection_failure_lifecycle_events(&test_context, 6);
-
-    ASSERT_SUCCESS(aws_mqtt5_client_stop(client, NULL, NULL));
-
-    s_wait_for_stopped_lifecycle_event(&test_context);
-
-    ASSERT_SUCCESS(s_verify_reconnection_exponential_backoff_timestamps(&test_context));
-
-    /* 6 (connecting, mqtt_connect, channel_shutdown, pending_reconnect) tuples (minus the final pending_reconnect) */
-    enum aws_mqtt5_client_state expected_states[] = {
-        AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN, AWS_MCS_PENDING_RECONNECT,
-        AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN, AWS_MCS_PENDING_RECONNECT,
-        AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN, AWS_MCS_PENDING_RECONNECT,
-        AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN, AWS_MCS_PENDING_RECONNECT,
-        AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN, AWS_MCS_PENDING_RECONNECT,
-        AWS_MCS_CONNECTING, AWS_MCS_MQTT_CONNECT, AWS_MCS_CHANNEL_SHUTDOWN,
-    };
-    ASSERT_SUCCESS(s_verify_client_state_sequence(&test_context, expected_states, AWS_ARRAY_SIZE(expected_states)));
-
-    aws_mqtt_client_connection_release(adapter);
-
-    aws_mqtt5_client_mock_test_fixture_clean_up(&test_context);
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_set_reconnect_timeout, s_mqtt3to5_adapter_set_reconnect_timeout_fn)
-
-static int s_mqtt3to5_adapter_connect_success_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture fixture;
-    ASSERT_SUCCESS(aws_mqtt3_to_mqtt5_adapter_test_fixture_init(&fixture, allocator, &test_fixture_options, NULL));
-
-    struct aws_mqtt_client_connection *adapter = fixture.connection;
-
-    struct aws_mqtt_connection_options connection_options;
-    s_init_adapter_connection_options_from_fixture(&connection_options, &fixture);
-
-    connection_options.on_connection_complete = s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_connection_complete;
-    connection_options.user_data = &fixture;
-
-    aws_mqtt_client_connection_connect(adapter, &connection_options);
-
-    s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_CONNECTION_COMPLETE, 1);
-
-    struct aws_mqtt3_lifecycle_event expected_events[] = {
-        {
-            .type = AWS_MQTT3_LET_CONNECTION_COMPLETE,
-        },
-    };
-    s_aws_mqtt3_to_mqtt5_adapter_test_fixture_verify_lifecycle_sequence(
-        &fixture, AWS_ARRAY_SIZE(expected_events), expected_events);
-
-    aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(&fixture);
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_connect_success, s_mqtt3to5_adapter_connect_success_fn)
-
-static int s_do_mqtt3to5_adapter_connect_success_disconnect_success_cycle(
-    struct aws_allocator *allocator,
-    size_t iterations) {
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture fixture;
-    ASSERT_SUCCESS(aws_mqtt3_to_mqtt5_adapter_test_fixture_init(&fixture, allocator, &test_fixture_options, NULL));
-
-    struct aws_mqtt_client_connection *adapter = fixture.connection;
-
-    for (size_t i = 0; i < iterations; ++i) {
-        struct aws_mqtt_connection_options connection_options;
-        s_init_adapter_connection_options_from_fixture(&connection_options, &fixture);
-
-        connection_options.on_connection_complete =
-            s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_connection_complete;
-        connection_options.user_data = &fixture;
-
-        aws_mqtt_client_connection_connect(adapter, &connection_options);
-
-        s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_CONNECTION_COMPLETE, i + 1);
-
-        aws_mqtt_client_connection_disconnect(
-            adapter, s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_disconnection_complete, &fixture);
-
-        s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_DISCONNECTION_COMPLETE, i + 1);
-        s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_CLOSED, i + 1);
-
-        struct aws_mqtt3_lifecycle_event expected_event_sequence[] = {
-            {
-                .type = AWS_MQTT3_LET_CONNECTION_COMPLETE,
-            },
-            {
-                .type = AWS_MQTT3_LET_DISCONNECTION_COMPLETE,
-            },
-            {
-                .type = AWS_MQTT3_LET_CLOSED,
-            },
-        };
-
-        size_t expected_event_count = (i + 1) * 3;
-        struct aws_mqtt3_lifecycle_event *expected_events =
-            aws_mem_calloc(allocator, expected_event_count, sizeof(struct aws_mqtt3_lifecycle_event));
-        for (size_t j = 0; j < i + 1; ++j) {
-            for (size_t k = 0; k < 3; ++k) {
-                *(expected_events + j * 3 + k) = expected_event_sequence[k];
-            }
-        }
-
-        s_aws_mqtt3_to_mqtt5_adapter_test_fixture_verify_lifecycle_sequence(
-            &fixture, expected_event_count, expected_events);
-
-        aws_mem_release(allocator, expected_events);
-    }
-
-    aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(&fixture);
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-static int s_mqtt3to5_adapter_connect_success_disconnect_success_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    ASSERT_SUCCESS(s_do_mqtt3to5_adapter_connect_success_disconnect_success_cycle(allocator, 1));
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(
-    mqtt3to5_adapter_connect_success_disconnect_success,
-    s_mqtt3to5_adapter_connect_success_disconnect_success_fn)
-
-static int s_mqtt3to5_adapter_connect_success_disconnect_success_thrice_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    ASSERT_SUCCESS(s_do_mqtt3to5_adapter_connect_success_disconnect_success_cycle(allocator, 3));
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(
-    mqtt3to5_adapter_connect_success_disconnect_success_thrice,
-    s_mqtt3to5_adapter_connect_success_disconnect_success_thrice_fn)
-
-static int s_mqtt3to5_adapter_connect_success_connect_failure_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture fixture;
-    ASSERT_SUCCESS(aws_mqtt3_to_mqtt5_adapter_test_fixture_init(&fixture, allocator, &test_fixture_options, NULL));
-
-    struct aws_mqtt_client_connection *adapter = fixture.connection;
-
-    struct aws_mqtt_connection_options connection_options;
-    s_init_adapter_connection_options_from_fixture(&connection_options, &fixture);
-
-    connection_options.on_connection_complete = s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_connection_complete;
-    connection_options.user_data = &fixture;
-
-    aws_mqtt_client_connection_connect(adapter, &connection_options);
-
-    s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_CONNECTION_COMPLETE, 1);
-
-    aws_mqtt_client_connection_connect(adapter, &connection_options);
-
-    s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_CONNECTION_COMPLETE, 2);
-
-    struct aws_mqtt3_lifecycle_event expected_events[] = {
-        {
-            .type = AWS_MQTT3_LET_CONNECTION_COMPLETE,
-        },
-        {
-            .type = AWS_MQTT3_LET_CONNECTION_COMPLETE,
-            .error_code = AWS_ERROR_MQTT_ALREADY_CONNECTED,
-        },
-    };
-    s_aws_mqtt3_to_mqtt5_adapter_test_fixture_verify_lifecycle_sequence(
-        &fixture, AWS_ARRAY_SIZE(expected_events), expected_events);
-
-    aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(&fixture);
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_connect_success_connect_failure, s_mqtt3to5_adapter_connect_success_connect_failure_fn)
-
-static int s_mqtt3to5_adapter_connect_success_sloppy_shutdown_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture fixture;
-    ASSERT_SUCCESS(aws_mqtt3_to_mqtt5_adapter_test_fixture_init(&fixture, allocator, &test_fixture_options, NULL));
-
-    struct aws_mqtt_client_connection *adapter = fixture.connection;
-
-    struct aws_mqtt_connection_options connection_options;
-    s_init_adapter_connection_options_from_fixture(&connection_options, &fixture);
-
-    connection_options.on_connection_complete = s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_connection_complete;
-    connection_options.user_data = &fixture;
-
-    aws_mqtt_client_connection_connect(adapter, &connection_options);
-
-    aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(&fixture);
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_connect_success_sloppy_shutdown, s_mqtt3to5_adapter_connect_success_sloppy_shutdown_fn)
-
-static void s_shutdown_server_channel_task_fn(struct aws_task *task, void *arg, enum aws_task_status status) {
-    struct aws_mqtt5_client *client = arg;
-
-    if (status == AWS_TASK_STATUS_CANCELED) {
-        goto done;
-    }
-
-    if (client->current_state == AWS_MCS_CONNECTED) {
-        aws_channel_shutdown(client->slot->channel, AWS_ERROR_MQTT5_CLIENT_TERMINATED);
-    }
-
-done:
-
-    aws_mem_release(client->allocator, task);
-
-    aws_mqtt5_client_release(client);
-}
-
-static int s_aws_mqtt5_server_disconnect_after_connect(
-    void *packet,
-    struct aws_mqtt5_server_mock_connection_context *connection,
-    void *user_data) {
-    int result = s_aws_mqtt5_mock_server_handle_connect_always_succeed(packet, connection, user_data);
-
-    struct aws_mqtt5_client *client = connection->test_fixture->client;
-    aws_mqtt5_client_acquire(client);
-
-    struct aws_task *shutdown_task = aws_mem_calloc(connection->allocator, 1, sizeof(struct aws_task));
-    aws_task_init(shutdown_task, s_shutdown_server_channel_task_fn, client, "ShutdownMockServerChannel");
-
-    uint64_t now = 0;
-    aws_high_res_clock_get_ticks(&now);
-    uint64_t scheduled_time = now + aws_timestamp_convert(1, AWS_TIMESTAMP_SECS, AWS_TIMESTAMP_NANOS, NULL);
-
-    aws_event_loop_schedule_task_future(connection->test_fixture->client->loop, shutdown_task, scheduled_time);
-
-    return result;
-}
-
-static int s_mqtt3to5_adapter_connect_bad_connectivity_fn(struct aws_allocator *allocator, void *ctx) {
-    (void)ctx;
-
-    aws_mqtt_library_init(allocator);
-
-    struct mqtt5_client_test_options test_options;
-    s_mqtt5_client_test_init_default_options(&test_options);
-
-    test_options.server_function_table.packet_handlers[AWS_MQTT5_PT_CONNECT] =
-        s_aws_mqtt5_server_disconnect_after_connect;
-
-    struct aws_mqtt5_client_mqtt5_mock_test_fixture_options test_fixture_options = {
-        .client_options = &test_options.client_options,
-        .server_function_table = &test_options.server_function_table,
-    };
-
-    struct aws_mqtt3_to_mqtt5_adapter_test_fixture fixture;
-    ASSERT_SUCCESS(aws_mqtt3_to_mqtt5_adapter_test_fixture_init(&fixture, allocator, &test_fixture_options, NULL));
-
-    struct aws_mqtt_client_connection *adapter = fixture.connection;
-
-    struct aws_mqtt_connection_options connection_options;
-    s_init_adapter_connection_options_from_fixture(&connection_options, &fixture);
-
-    connection_options.on_connection_complete = s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_connection_complete;
-    connection_options.user_data = &fixture;
-
-    aws_mqtt_client_connection_connect(adapter, &connection_options);
-
-    s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_RESUMED, 3);
-
-    aws_mqtt_client_connection_disconnect(
-        adapter, s_aws_mqtt3_to_mqtt5_adapter_test_fixture_record_disconnection_complete, &fixture);
-
-    s_wait_for_n_adapter_lifecycle_events(&fixture, AWS_MQTT3_LET_CLOSED, 1);
-
-    struct aws_mqtt3_lifecycle_event expected_events[] = {
-        {
-            .type = AWS_MQTT3_LET_CONNECTION_COMPLETE,
-        },
-        {
-            .type = AWS_MQTT3_LET_INTERRUPTED,
-            .error_code = AWS_ERROR_MQTT5_CLIENT_TERMINATED,
-        },
-        {
-            .type = AWS_MQTT3_LET_RESUMED,
-        },
-        {
-            .type = AWS_MQTT3_LET_INTERRUPTED,
-            .error_code = AWS_ERROR_MQTT5_CLIENT_TERMINATED,
-        },
-        {
-            .type = AWS_MQTT3_LET_RESUMED,
-        },
-        {
-            .type = AWS_MQTT3_LET_INTERRUPTED,
-            .error_code = AWS_ERROR_MQTT5_CLIENT_TERMINATED,
-        },
-        {
-            .type = AWS_MQTT3_LET_RESUMED,
-        },
-        {
-            .type = AWS_MQTT3_LET_DISCONNECTION_COMPLETE,
-        },
-        {
-            .type = AWS_MQTT3_LET_CLOSED,
-        },
-    };
-    s_aws_mqtt3_to_mqtt5_adapter_test_fixture_verify_lifecycle_sequence(
-        &fixture, AWS_ARRAY_SIZE(expected_events), expected_events);
-
-    aws_mqtt3_to_mqtt5_adapter_test_fixture_clean_up(&fixture);
-    aws_mqtt_library_clean_up();
-
-    return AWS_OP_SUCCESS;
-}
-
-AWS_TEST_CASE(mqtt3to5_adapter_connect_bad_connectivity, s_mqtt3to5_adapter_connect_bad_connectivity_fn)
